Add scoreWatcher method calls to main game code.
Added the scoreWatcher method calls to the main game code. Also tested
the game to make sure that the code actually worked and displayed the
scores correctly. Unfortunately had to make the scores global variables,
which I'm pretty sure is frowned upon, but making them local would have taken
a few more minutes to do. I figure it does not matter for the purposes of
this exercise.